requirements.txt
Django==2.0.5
django-enumfields
celery==4.1.0
django-semanticui-forms